# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SERMACROPS EDI Process Flow Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateCompany
      tags: [companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    delete:
      operationId: deleteCompany
      tags: [companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /edi-documents:
    get:
      operationId: listEdiDocuments
      tags: [ediDocuments]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: direction
          in: query
          schema:
            type: string
        - name: documentType
          in: query
          schema:
            type: string
        - name: companyId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EdiDocument"
    post:
      operationId: createEdiDocument
      tags: [ediDocuments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EdiDocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EdiDocument"
  /edi-documents/{id}:
    get:
      operationId: getEdiDocument
      tags: [ediDocuments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EdiDocument"
    patch:
      operationId: updateEdiDocument
      tags: [ediDocuments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EdiDocumentUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EdiDocument"
    delete:
      operationId: deleteEdiDocument
      tags: [ediDocuments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /edi-documents/{id}/send:
    post:
      operationId: sendEdiDocument
      tags: [ediDocuments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendResult"
  /edi-documents/{id}/preview:
    get:
      operationId: previewEdiDocument
      tags: [ediDocuments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/X12Preview"
  /edi/inbound:
    post:
      operationId: receiveInbound
      tags: [inbound]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InboundPayload"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboundReceipt"
  /inbound-messages:
    get:
      operationId: listInboundMessages
      tags: [inbound]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: documentType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InboundMessage"
  /inbound-messages/{id}:
    get:
      operationId: getInboundMessage
      tags: [inbound]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboundMessage"
...
components:
  schemas:
...
    Company:
      type: object
      required: [id, name, ediId, type, isActive, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        ediId:
          type: string
        type:
          type: string
        address:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    CompanyInput:
      type: object
      required: [name, ediId, type]
      properties:
        name:
          type: string
        ediId:
          type: string
        type:
          type: string
        address:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        isActive:
          type: boolean
    CompanyUpdate:
      type: object
      properties:
        name:
          type: string
        ediId:
          type: string
        type:
          type: string
        address:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        isActive:
          type: boolean
...
    EdiDocument:
      type: object
      required: [id, documentType, direction, status, senderId, receiverId, controlNumber, retryCount, createdAt, updatedAt]
      properties:
        id:
          type: string
        documentType:
          type: string
        direction:
          type: string
        status:
          type: string
        senderId:
          type: string
        senderName:
          type: ["string", "null"]
        receiverId:
          type: string
        receiverName:
          type: ["string", "null"]
        controlNumber:
          type: string
        referenceNumber:
          type: ["string", "null"]
        poNumber:
          type: ["string", "null"]
        shipDate:
          type: ["string", "null"]
        deliveryDate:
          type: ["string", "null"]
        totalAmount:
          type: ["number", "null"]
        lineItems:
          type: ["string", "null"]
        paymentTerms:
          type: ["string", "null"]
        shippingDetails:
          type: ["string", "null"]
        x12Content:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        retryCount:
          type: integer
        lastResponseCode:
          type: ["integer", "null"]
        lastResponseBody:
          type: ["string", "null"]
        sentAt:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
        transactionId:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    EdiDocumentInput:
      type: object
      required: [documentType, direction, senderId, receiverId]
      properties:
        documentType:
          type: string
        direction:
          type: string
        senderId:
          type: string
        receiverId:
          type: string
        referenceNumber:
          type: string
        poNumber:
          type: string
        shipDate:
          type: string
        deliveryDate:
          type: string
        totalAmount:
          type: number
        lineItems:
          type: string
        paymentTerms:
          type: string
        shippingDetails:
          type: string
        notes:
          type: string
        transactionId:
          type: string
        status:
          type: string
    EdiDocumentUpdate:
      type: object
      properties:
        documentType:
          type: string
        status:
          type: string
        referenceNumber:
          type: string
        poNumber:
          type: string
        shipDate:
          type: string
        deliveryDate:
          type: string
        totalAmount:
          type: number
        lineItems:
          type: string
        paymentTerms:
          type: string
        shippingDetails:
          type: string
        notes:
          type: string
        x12Content:
          type: string
    SendResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        responseCode:
          type: ["integer", "null"]
        responseBody:
          type: ["string", "null"]
        sentAt:
          type: ["string", "null"]
    X12Preview:
      type: object
      required: [content, documentType]
      properties:
        content:
          type: string
        documentType:
          type: string
    InboundPayload:
      type: object
      required: [x12Content]
      properties:
        x12Content:
          type: string
    InboundReceipt:
      type: object
      required: [success, messageId]
      properties:
        success:
          type: boolean
        messageId:
          type: string
        documentType:
          type: ["string", "null"]
        sender:
          type: ["string", "null"]
        receiver:
          type: ["string", "null"]
        errors:
          type: array
          items:
            type: string
    InboundMessage:
      type: object
      required: [id, rawPayload, status, createdAt]
      properties:
        id:
          type: string
        documentType:
          type: ["string", "null"]
        senderId:
          type: ["string", "null"]
        senderName:
          type: ["string", "null"]
        receiverId:
          type: ["string", "null"]
        receiverName:
          type: ["string", "null"]
        rawPayload:
          type: string
        parsedData:
          type: ["string", "null"]
        status:
          type: string
        validationErrors:
          type: ["string", "null"]
        controlNumber:
          type: ["string", "null"]
        createdAt:
          type: string
        processedAt:
          type: ["string", "null"]
...