# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SERMACROPS EDI Process Flow Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: string;
  name: string;
  ediId: string;
  type: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  isActive: boolean;
  createdAt: string;
}